Geneformer/docs/requirements.txt
.
sphinx_rtd_theme==2.0.0
nbsphinx==0.9.3
